[ui] Move AutomaterializeMiddlePanelWithData
