+++ js_modules/dagster-ui/packages/ui-core/src/assets/AutoMaterializePolicyPage/AutomaterializeMiddlePanel.tsx
@@ -1,36 +1,12 @@
-import {gql, useQuery} from '@apollo/client';
-import {
-  BaseTag,
-  Body2,
-  Box,
-  Colors,
-  Icon,
-  MenuItem,
-  MiddleTruncate,
-  NonIdealState,
-  Popover,
-  Subheading,
-  Subtitle2,
-  Tag,
-  TagSelectorContainer,
-  TagSelectorDefaultTagTooltipStyle,
-  TagSelectorWithSearch,
-} from '@dagster-io/ui-components';
+import {useQuery} from '@apollo/client';
+import {Body2, Box, NonIdealState, Subheading} from '@dagster-io/ui-components';
 import React from 'react';
-import styled from 'styled-components';
 
-import {StatusDot} from './AutomaterializeLeftPanel';
-import {AutomaterializeRunsTable} from './AutomaterializeRunsTable';
+import {AutomaterializeMiddlePanelWithData} from './AutomaterializeMiddlePanelWithData';
 import {
   GET_EVALUATIONS_QUERY,
   GET_EVALUATIONS_SPECIFIC_PARTITION_QUERY,
 } from './GetEvaluationsQuery';
-import {PartitionSubsetList} from './PartitionSegmentWithPopover';
-import {PolicyEvaluationTable} from './PolicyEvaluationTable';
-import {
-  FullPartitionsQuery,
-  FullPartitionsQueryVariables,
-} from './types/AutomaterializeMiddlePanel.types';
 import {
   AssetConditionEvaluationRecordFragment,
   GetEvaluationsQuery,
@@ -39,9 +15,7 @@ import {
   GetEvaluationsSpecificPartitionQueryVariables,
 } from './types/GetEvaluationsQuery.types';
 import {ErrorWrapper} from '../../app/PythonErrorInfo';
-import {formatElapsedTimeWithMsec} from '../../app/Util';
-import {Timestamp} from '../../app/time/Timestamp';
-import {DimensionPartitionKeys, SensorType} from '../../graphql/types';
+import {SensorType} from '../../graphql/types';
 import {useQueryPersistedState} from '../../hooks/useQueryPersistedState';
 import {useBlockTraceOnQueryResult} from '../../performance/TraceContext';
 import {AnchorButton} from '../../ui/AnchorButton';
@@ -56,8 +30,6 @@ interface Props {
   selectedEvaluation?: AssetConditionEvaluationRecordFragment;
   definition?: AssetViewDefinitionNodeFragment | null;
 }
-
-const emptyArray: any[] = [];
 
 export const SELECTED_PARTITION_QUERY_STRING_KEY = 'selectedPartition';
 
@@ -217,265 +189,3 @@ export const AutomaterializeMiddlePanel = (props: Props) => {
     />
   );
 };
-
-export const AutomaterializeMiddlePanelWithData = ({
-  selectedEvaluation,
-  definition,
-  selectPartition,
-  specificPartitionData,
-  selectedPartition,
-}: {
-  definition?: AssetViewDefinitionNodeFragment | null;
-  selectedEvaluation?: AssetConditionEvaluationRecordFragment;
-  selectPartition: (partitionKey: string | null) => void;
-  specificPartitionData?: GetEvaluationsSpecificPartitionQuery;
-  selectedPartition: string | null;
-}) => {
-  const evaluation = selectedEvaluation?.evaluation;
-  const rootEvaluationNode = React.useMemo(
-    () => evaluation?.evaluationNodes.find((node) => node.uniqueId === evaluation.rootUniqueId),
-    [evaluation],
-  );
-  const rootPartitionedEvaluationNode =
-    rootEvaluationNode?.__typename === 'PartitionedAssetConditionEvaluationNode'
-      ? rootEvaluationNode
-      : null;
-
-  const statusTag = React.useMemo(() => {
-    if (selectedEvaluation?.numRequested) {
-      if (definition?.partitionDefinition) {
-        return (
-          <Box flex={{direction: 'row', gap: 4, alignItems: 'center'}}>
-            <Popover
-              interactionKind="hover"
-              placement="bottom"
-              hoverOpenDelay={50}
-              hoverCloseDelay={50}
-              content={
-                <PartitionSubsetList
-                  description="Requested assets"
-                  subset={rootPartitionedEvaluationNode!.trueSubset}
-                  selectPartition={selectPartition}
-                />
-              }
-            >
-              <Tag intent="success">
-                <Box flex={{direction: 'row', gap: 4, alignItems: 'center'}}>
-                  <StatusDot $color={Colors.accentGreen()} $size={8} />
-                  {selectedEvaluation.numRequested} Requested
-                </Box>
-              </Tag>
-            </Popover>
-            {selectedEvaluation.numRequested === 1 ? (
-              <Tag icon="partition">
-                {rootPartitionedEvaluationNode!.trueSubset.subsetValue.partitionKeys![0]}
-              </Tag>
-            ) : null}
-          </Box>
-        );
-      }
-      return (
-        <Tag intent="success">
-          <Box flex={{direction: 'row', gap: 4, alignItems: 'center'}}>
-            <StatusDot $color={Colors.accentGreen()} />
-            Requested
-          </Box>
-        </Tag>
-      );
-    }
-    return (
-      <Tag>
-        <Box flex={{direction: 'row', gap: 4, alignItems: 'center'}}>
-          <StatusDot $color={Colors.accentGray()} />
-          Not Requested
-        </Box>
-      </Tag>
-    );
-  }, [
-    definition?.partitionDefinition,
-    rootPartitionedEvaluationNode,
-    selectPartition,
-    selectedEvaluation?.numRequested,
-  ]);
-
-  const fullPartitionsQueryResult = useQuery<FullPartitionsQuery, FullPartitionsQueryVariables>(
-    FULL_PARTITIONS_QUERY,
-    {
-      variables: definition
-        ? {
-            assetKey: {path: definition.assetKey.path},
-          }
-        : undefined,
-      skip: !definition?.assetKey,
-    },
-  );
-  useBlockTraceOnQueryResult(fullPartitionsQueryResult, 'FullPartitionsQuery', {
-    skip: !definition?.assetKey,
-  });
-
-  const {data} = fullPartitionsQueryResult;
-
-  let partitionKeys: DimensionPartitionKeys[] = emptyArray;
-  if (data?.assetNodeOrError.__typename === 'AssetNode') {
-    partitionKeys = data.assetNodeOrError.partitionKeysByDimension;
-  }
-
-  const allPartitions = React.useMemo(() => {
-    if (partitionKeys.length === 1) {
-      return partitionKeys[0]!.partitionKeys;
-    } else if (partitionKeys.length === 2) {
-      const firstSet = partitionKeys[0]!.partitionKeys;
-      const secondSet = partitionKeys[1]!.partitionKeys;
-      return firstSet.flatMap((key1) => secondSet.map((key2) => `${key1}|${key2}`));
-    } else if (partitionKeys.length > 2) {
-      throw new Error('Only 2 dimensions are supported');
-    }
-    return [];
-  }, [partitionKeys]);
-
-  return (
-    <Box flex={{direction: 'column', grow: 1}}>
-      <Box
-        style={{flex: '0 0 48px'}}
-        padding={{horizontal: 16}}
-        border="bottom"
-        flex={{alignItems: 'center', justifyContent: 'space-between'}}
-      >
-        <Subheading>Result</Subheading>
-      </Box>
-      {selectedEvaluation ? (
-        <Box padding={{horizontal: 24, vertical: 12}}>
-          <Box border="bottom" padding={{vertical: 12}} margin={{bottom: 12}}>
-            <div style={{display: 'grid', gridTemplateColumns: '1fr 1fr 1fr', gap: 24}}>
-              <Box flex={{direction: 'column', gap: 5}}>
-                <Subtitle2>Evaluation result</Subtitle2>
-                <div>{statusTag}</div>
-              </Box>
-              {selectedEvaluation?.timestamp ? (
-                <Box flex={{direction: 'column', gap: 5}}>
-                  <Subtitle2>Timestamp</Subtitle2>
-                  <Timestamp timestamp={{unix: selectedEvaluation?.timestamp}} />
-                </Box>
-              ) : null}
-              <Box flex={{direction: 'column', gap: 5}}>
-                <Subtitle2>Duration</Subtitle2>
-                <div>
-                  {selectedEvaluation?.startTimestamp && selectedEvaluation?.endTimestamp
-                    ? formatElapsedTimeWithMsec(
-                        (selectedEvaluation.endTimestamp - selectedEvaluation.startTimestamp) *
-                          1000,
-                      )
-                    : '\u2013'}
-                </div>
-              </Box>
-            </div>
-          </Box>
-          <Box border="bottom" padding={{vertical: 12}} margin={{vertical: 12}}>
-            <Subtitle2>Runs launched ({selectedEvaluation.runIds.length})</Subtitle2>
-          </Box>
-          <AutomaterializeRunsTable runIds={selectedEvaluation.runIds} />
-          <Box border="bottom" padding={{vertical: 12}}>
-            <Subtitle2>Policy evaluation</Subtitle2>
-          </Box>
-          {definition?.partitionDefinition ? (
-            <Box padding={{vertical: 12}} flex={{justifyContent: 'flex-end'}}>
-              <TagSelectorWrapper>
-                <TagSelectorWithSearch
-                  closeOnSelect
-                  placeholder="Select a partition to view its result"
-                  allTags={allPartitions}
-                  selectedTags={selectedPartition ? [selectedPartition] : []}
-                  setSelectedTags={(tags) => {
-                    selectPartition(tags[tags.length - 1] || null);
-                  }}
-                  renderDropdownItem={(tag, props) => (
-                    <MenuItem text={tag} onClick={props.toggle} />
-                  )}
-                  renderDropdown={(dropdown) => (
-                    <Box padding={{top: 8, horizontal: 4}} style={{width: '370px'}}>
-                      {dropdown}
-                    </Box>
-                  )}
-                  renderTag={(tag, tagProps) => (
-                    <BaseTag
-                      key={tag}
-                      textColor={Colors.textLight()}
-                      fillColor={Colors.backgroundGray()}
-                      icon={<Icon name="partition" color={Colors.accentGray()} />}
-                      label={
-                        <div
-                          style={{
-                            display: 'grid',
-                            gridTemplateColumns: '1fr auto',
-                            gap: 4,
-                            justifyContent: 'space-between',
-                            alignItems: 'center',
-                            maxWidth: '120px',
-                          }}
-                          data-tooltip={tag}
-                          data-tooltip-style={TagSelectorDefaultTagTooltipStyle}
-                        >
-                          <MiddleTruncate text={tag} />
-                          <Box style={{cursor: 'pointer'}} onClick={tagProps.remove}>
-                            <Icon name="close" />
-                          </Box>
-                        </div>
-                      }
-                    />
-                  )}
-                  usePortal={false}
-                />
-                <SearchIconWrapper>
-                  <Icon name="search" />
-                </SearchIconWrapper>
-              </TagSelectorWrapper>
-            </Box>
-          ) : null}
-          <PolicyEvaluationTable
-            evaluationRecord={
-              selectedPartition && specificPartitionData?.assetConditionEvaluationForPartition
-                ? {evaluation: specificPartitionData.assetConditionEvaluationForPartition}
-                : selectedEvaluation
-            }
-            definition={definition}
-            selectPartition={selectPartition}
-          />
-        </Box>
-      ) : null}
-    </Box>
-  );
-};
-
-const FULL_PARTITIONS_QUERY = gql`
-  query FullPartitionsQuery($assetKey: AssetKeyInput!) {
-    assetNodeOrError(assetKey: $assetKey) {
-      ... on AssetNode {
-        id
-        partitionKeysByDimension {
-          name
-          type
-          partitionKeys
-        }
-      }
-    }
-  }
-`;
-const TagSelectorWrapper = styled.div`
-  position: relative;
-
-  ${TagSelectorContainer} {
-    width: 370px;
-    padding-left: 32px;
-    height: 36px;
-  }
-`;
-
-const SearchIconWrapper = styled.div`
-  position: absolute;
-  left: 12px;
-  top: 0px;
-  bottom: 0px;
-  pointer-events: none;
-  display: flex;
-  align-items: center;
-`;
